Work on cleaning up project and have things better split up.
