+++ backend/app/main.py
@@ -1,12 +1,12 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from backend.routes.login import login_router
-from backend.routes.reports import report_router
-from backend.routes.users import user_router
+from app.routes.login import login_router
+from app.routes.reports import report_router
+from app.routes.users import user_router
 
-from backend.database import engine
-from backend.models import Base
+from app.database import engine
+from app.models import Base
 
 app = FastAPI()
