+++ test.yml
@@ -11,16 +11,16 @@ services:
     ports:
       - "5432:5432"
     healthcheck:
-      test: [ "CMD-SHELL", "pg_isready -U user -d test" ]
+      test: [ "CMD-SHELL", "pg_isready -U testuser -d test" ]
       interval: 10s
       timeout: 5s
       retries: 5
 
   backend:
     build:
-      context: .
+      context: backend
       dockerfile: test.Dockerfile
-    working_dir: /src/backend
+    working_dir: /backend/app
     hostname: backend
 #    ports:
 #      - "8000:8000"
